+++ SnowGL/src/GPU_Mesh.cpp
@@ -17,9 +17,9 @@ namespace SnowGL
 		m_VAO->addBuffer(*m_VBO, _mesh.layout);
 
 		// load data to VBOs
-		m_VBO->loadData(_mesh.vertices.data(), _mesh.layout.getStride() * _mesh.vertices.size());
-		m_IBO->loadData(_mesh.indices.data(), (GLuint)_mesh.indices.size(), sizeof(unsigned int) * _mesh.indices.size());
+		m_VBO->loadData(_mesh.vertices.data(), _mesh.layout.getStride() * (GLuint)_mesh.vertices.size());
+		m_IBO->loadData(_mesh.indices.data(), (GLuint)_mesh.indices.size(), sizeof(unsigned int) * (GLuint)_mesh.indices.size());
 
-		m_vertexCount = _mesh.indices.size();
+		m_vertexCount = (int)_mesh.indices.size();
 	}
 }
